+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - correctly placing skill messages after the user prompt when used from custom commands
 - added skillsToActivate to onAgentStarted extension event to allow activating skills
 - added noProxy support to network settings
+- replaced floating placement with task, project, and app-level floating placements in UI extensions
 
 ## [0.70.0]